start adding actual ui
