+++ src/components/index.js
@@ -14,5 +14,5 @@ export { default as PokeTokenLaunchPage} from "./pokemon/token_launch_page";
 export { default as Rewards} from "./apps/rewards";
 export { default as Verified} from "./apps/verified";
 export { default as IceCreamPost} from "./blog/posts/8_IceCream/ice_cream";
-export { default as Dungeon} from "./apps/dungeon";
+export { default as Dungeon} from "./apps/dungeon/dungeon";
